requirements.txt
paho-mqtt~=1.6.1
setuptools~=60.9.3
